relocates topic specific functions
